feat(movieList): 더미 데이터와 상영 정보 바인딩

+++ app/src/main/java/woowacourse/movie/presentation/movieList/MovieListClickListener.kt
@@ -1,5 +1,5 @@
 package woowacourse.movie.presentation.movieList
 
 interface MovieListClickListener {
-    fun ticketingButtonClick(movieId: Int)
+    fun ticketingButtonClick(movieId: Long)
 }
