+++ app/src/main/java/woowacourse/movie/presentation/ticketing/TicketingPresenter.kt
@@ -17,7 +17,7 @@ class TicketingPresenter(
     private lateinit var selectedMovie: Movie
     private var selectedTimePosition = 0
 
-    override fun loadMovieData(id: Int) {
+    override fun loadMovieData(id: Long) {
         movieRepository.findMovieById(id)
             .onSuccess { movie ->
                 selectedMovie = movie
